improvement(domains): change from GoDaddy to Domainr for domain sourcing
Change from GoDaddy API to Domainr for checking domain statuses.
