+++ dev.env
@@ -4,3 +4,4 @@ GODADDY_AUTH_TOKEN=
 TWITCH_CLIENT_ID=
 TWITTER_BEARER_TOKEN=
 YOUTUBE_API_KEY=
+RAPIDAPI_KEY=
